feat: implement useKeyboard hook to manage keyboard visibility and height
